backend/requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==22.10.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastapi==0.95.0
Fiona==1.9.2
GeoAlchemy2==0.13.1
geopandas==0.12.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
msgpack==1.0.3
munch==2.5.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.14.0
pynvim==0.4.3
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.0
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
wcwidth==0.2.6
